Adds the Popup Management
All pages can now call the openPopup(id) function to open a popup
with the given id.
The popup will be closed when the user clicks outside of it.

Add the ".outside-click-no-close" class to a popup element to prevent
it from closing when the user clicks outside of it.

Add ".open-popup" class to the button that opens the popup.
(otherwise it closes right away)